backend/requirements.txt
Pillow
torch == 1.10.0
torchvision == 0.11.1
matplotlib
numpy
mlflow
dagshub
fastapi
boto3
uvicorn
python-multipart